requirements.txt
python-nmap==0.7.1
requests==2.31.0
paramiko
h2o
pandas
numpy
jinja2
pytest
pytest-html
paho-mqtt>=1.6
telnetlib3
matplotlib
aiocoap
pymodbus==2.5.3
pyyaml
scikit-learn
anthropic
openai
google-genai
